api_multi_model/requirements.txt
Django
djangorestframework
coreapi
pygments
psycopg2
djangorestframework-jwt
markdown
pillow
